feat: rehaul how commands are built, break the database, general clean up
