+++ src/commands/implementations/fun/hug.ts
@@ -1,19 +1,19 @@
 import type ICommand from "../../ICommand.ts";
-import ExtendedSlashCommandBuilder from "../../ExtendedSlashCommandBuilder.ts";
+import CommandBuilder from "../../CommandBuilder.ts";
 import {ChatInputCommandInteraction} from "discord.js";
 import {randomFeeling} from "../../../utils/feelings.ts";
 
-export default class SlapCommand implements ICommand {
-    name: string = "slap";
-    description: string = "Slaps your victim";
+export default class Hug implements ICommand {
+    name: string = "hug";
+    description: string = "Give someone a big fat cuddly wuddly";
 
     getCommand() {
-        return new ExtendedSlashCommandBuilder()
+        return new CommandBuilder()
             .setName(this.name)
             .setDescription(this.description)
             .addUserOption(option => option
                 .setName("target")
-                .setDescription("The user to slap")
+                .setDescription("The user to hug")
                 .setRequired(true)
             );
     }
@@ -30,7 +30,7 @@ export default class SlapCommand implements ICommand {
                 : "messages";
 
         // Get a random feeling message based on the determined group
-        const message = randomFeeling("SLAP", group)
+        const message = randomFeeling("HUG", group)
             .replace("{user}", interaction.user.toString())
             .replace("{target}", target.toString());
