+++ src/commands/implementations/fun/slap.ts
@@ -1,19 +1,19 @@
 import type ICommand from "../../ICommand.ts";
-import ExtendedSlashCommandBuilder from "../../ExtendedSlashCommandBuilder.ts";
+import CommandBuilder from "../../CommandBuilder.ts";
 import {ChatInputCommandInteraction} from "discord.js";
 import {randomFeeling} from "../../../utils/feelings.ts";
 
-export default class PokeCommand implements ICommand {
-    name: string = "poke";
-    description: string = "Pokes your victim";
+export default class Slap implements ICommand {
+    name: string = "slap";
+    description: string = "Slaps your victim";
 
     getCommand() {
-        return new ExtendedSlashCommandBuilder()
+        return new CommandBuilder()
             .setName(this.name)
             .setDescription(this.description)
             .addUserOption(option => option
                 .setName("target")
-                .setDescription("The user to poke")
+                .setDescription("The user to slap")
                 .setRequired(true)
             );
     }
@@ -30,7 +30,7 @@ export default class PokeCommand implements ICommand {
                 : "messages";
 
         // Get a random feeling message based on the determined group
-        const message = randomFeeling("POKE", group)
+        const message = randomFeeling("SLAP", group)
             .replace("{user}", interaction.user.toString())
             .replace("{target}", target.toString());
