+++ src/database/database.ts
@@ -1,5 +1,6 @@
 // TODO: Switch to postgres w/ drizzle
-// import { connect, disconnect, connection } from "mongoose";
+// import mongoose, { connect, disconnect } from "mongoose";
+// const {connection} = mongoose;
 //
 // export async function databaseConnect(mongoUri: string): Promise<void> {
 //   await connect(mongoUri);
@@ -12,6 +13,6 @@
 // export async function databaseLatency(): Promise<number> {
 //   const start = Date.now();
 //   if (!connection.db) throw new Error("No database connection");
-//   await connection.db.commands({ping: 1});
+//   await connection.db.command({ping: 1});
 //   return Date.now() - start;
 // }
